Drop SampleNtt() into its own file
